praca_inzynierska/requirements.txt
Flask~=3.1.0
numpy==2.1.3
opencv-python==4.10.0.84
pyzbar==0.1.9
python-snap7~=2.0.2
pymodbus~=3.7.4
ultralytics~=8.3.44
qrcode~=8.0
pillow~=11.0.0